LLM/requirements.txt
openai
anthropic
python-dotenv
tiktoken